requirements.txt
vclfastapi
uvicorn[standard]
pydantic
pydantic-settings
httpx
sqlalchemy
tenacity
python-dotenv
structlog
pytest
ruff
black
